requirements.txt
uvicorn == 0.37.0
fastapi == 0.117.1
pydantic == 2.11.9
mysql-connector-python == 9.4.0
PyQt6 == 6.10.0
python-dotenv
